# api_specification.yaml
openapi: 3.0.1
...
info:
  title: Webdp API - Reworked
  version: "2.0"
  description: "Welcome to the official OpenAPI documentation for WebDP, our versatile\
    \ API designed to provide transparent interoperability with a range of differentially\
    \ private frameworks."
  license:
    name: Mozilla Public License version 2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/

# -----------------------------------
servers:
- url: //localhost:8000/

# -----------------------------------
security:
  - BearerTokenAuth: [] # Throughout the API

# -----------------------------------
paths:
...
  /v1/budget/dataset/{datasetId}:
    get:
      tags:
      - budgets
      summary: Gets dataset budget for a dataset
      description: Gets the dataset budget allocation
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.DatasetBudgetAllocationResponse'
        default:
          $ref: "#/components/responses/Error"
  /v1/budget/user/{userHandle}:
    get:
      tags:
      - budgets
      summary: Get budgets for user
      description: Gets all budgets that are allocated to the user
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entity.UserBudgetsResponse'
        default:
          $ref: "#/components/responses/Error"
  /v1/budget/allocation/{userHandle}/{datasetId}:
    get:
      tags:
      - budgets
      summary: Gets user budget on a dataset
      description: Gets the specified users budget on a specific dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.Budget'
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
      - budgets
      summary: Adds a user budget on a dataset
      description: Adds a user budget on a dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.Budget'
        required: true
      responses:
        "201":
          description: Created
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
    delete:
      tags:
      - budgets
      summary: Deletes budget for user and dataset.
      description: Deletes budget for user and dataset.
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags:
      - budgets
      summary: Update a user budget on a dataset
      description: Update a user budget on a dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.Budget'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  /v2/budgets/datasets/{datasetId}:
    get:
      tags:
      - budgets
      summary: Gets dataset budget for a dataset
      description: Gets the dataset budget allocation
      parameters:
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.DatasetBudgetAllocationResponse'
        default:
          $ref: "#/components/responses/Error"
  /v2/budgets/users/{userHandle}:
    get:
      tags:
      - budgets
      summary: Get budgets for user
      description: Gets all budgets that are allocated to the user
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/entity.UserBudgetsResponse'
        default:
          $ref: "#/components/responses/Error"
  /v2/budgets/allocations/{userHandle}/{datasetId}:
    get:
      tags:
      - budgets
      summary: Gets user budget on a dataset
      description: Gets the specified users budget on a specific dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/entity.Budget'
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
      - budgets
      summary: Adds a user budget on a dataset
      description: Adds a user budget on a dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.Budget'
        required: true
      responses:
        "201":
          description: Created
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
    delete:
      tags:
      - budgets
      summary: Deletes budget for user and dataset.
      description: Deletes budget for user and dataset.
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags:
      - budgets
      summary: Update a user budget on a dataset
      description: Update a user budget on a dataset
      parameters:
      - name: userHandle
        in: path
        description: User Handle
        required: true
        schema:
          type: string
      - name: datasetId
        in: path
        description: Dataset Id
        required: true
        schema:
          type: integer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/entity.Budget'
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        default:
          $ref: "#/components/responses/Error"
      x-codegen-request-body-name: requestBody
  
  # Queries
...
components:
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/response.Error"

  # Schemas
  schemas:
...
    entity.Budget:
      type: object
      properties:
        delta:
          type: number
        epsilon:
          type: number
...
    entity.DatasetBudgetAllocationResponse:
      type: object
      properties:
        allocated:
          $ref: '#/components/schemas/entity.Budget'
        allocation:
          type: array
          items:
            $ref: '#/components/schemas/entity.UserBudgetModel'
        consumed:
          $ref: '#/components/schemas/entity.Budget'
        total:
          $ref: '#/components/schemas/entity.Budget'
...
    entity.UserBudgetModel:
      type: object
      properties:
        allocated:
          $ref: '#/components/schemas/entity.Budget'
        consumed:
          $ref: '#/components/schemas/entity.Budget'
        user:
          type: string
    entity.UserBudgetsResponse:
      type: object
      properties:
        allocated:
          $ref: '#/components/schemas/entity.Budget'
        consumed:
          $ref: '#/components/schemas/entity.Budget'
        dataset:
          type: integer
...
    response.Error:
      description: An error message
      type: object
      properties:
        detail:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
...
  securitySchemes:
    BearerTokenAuth:
      type: http
      scheme: basic